+++ NEWS.md
@@ -1,4 +1,10 @@
 ## 0.3
 
-* [#18](https://github.com/brodieG/validate/issues/18): better documentation for  NSE
-* [#11](https://github.com/brodieG/validate/issues/11): segfault when validating language objects
+* [#34](https://github.com/brodieG/validate/issues/34): allow substitution of
+  `.` symbol when part of `..`.
+* [#33](https://github.com/brodieG/validate/issues/33): prevent infinite
+  recursion with recursive symbol substitution
+* [#18](https://github.com/brodieG/validate/issues/18): better documentation for
+  NSE
+* [#11](https://github.com/brodieG/validate/issues/11): segfault when validating
+  language objects